requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asn1crypto==1.5.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cloud-sql-python-connector==1.18.3
cryptography==45.0.5
dnspython==2.7.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
multidict==6.6.3
pg8000==1.31.4
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.4
rsa==4.9.1
scramp==1.4.6
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
